added new admin endpoint

+++ routers/todos.py
@@ -8,6 +8,7 @@ from starlette import status
 from models import Todos
 from database import SessionLocal
 from .auth import get_current_user
+
 # engine,
 # from routers import auth
 
@@ -28,7 +29,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
-user_dependency= Annotated[dict, Depends(get_current_user)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -38,16 +39,15 @@ class TodoRequest(BaseModel):
     complete: bool
 
 
-
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(user:user_dependency, db: db_dependency):
+async def read_all(user: user_dependency, db: db_dependency):
     if user is None:
         raise HTTPException(status_code=401, detail='Authentication Failed')
     return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_by_id(user:user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+async def read_by_id(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
     if user is None:
         raise HTTPException(status_code=401, detail='Authentication Failed')
 
@@ -58,7 +58,7 @@ async def read_by_id(user:user_dependency, db: db_dependency, todo_id: int = Pat
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(user:user_dependency, db: db_dependency, todo_request: TodoRequest):
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
     if user is None:
         raise HTTPException(status_code=401, detail='Authentication Failed')
     todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id'))
@@ -68,7 +68,7 @@ async def create_todo(user:user_dependency, db: db_dependency, todo_request: Tod
 
 
 @router.put("/todo{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(user:user_dependency,db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+async def update_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
     if user is None:
         raise HTTPException(status_code=401, detail='Authentication Failed')
     todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
@@ -85,10 +85,12 @@ async def update_todo(user:user_dependency,db: db_dependency, todo_request: Todo
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found with provided id.')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
 
     db.commit()
